refactor: relocate DTOs into bounded-context domain packages
